Timer and Guess Interface
I added a timer than will run for the host. This will trigger a guess event after X time. Guess event is still being implemented. It will contact the server and call a method which will transport players to a location, show them a dialogue CS and then give them a timer to use the guess interface(which has been added) to submit their guess.
Host will have a method checking if all are submitted/time is up. whichever happens first will then trigger a check of answers. this check will lead to end of game or send players back to the map.

Next...
1. Keep working on guess implementation
2. NPC AI (movement and "knowledge"/dialogue AI)
3. Server integration
4. game over condition
5. menus for game start and options
6. refine strings and visual assets
7. Add information about the other players to the in game menu (in the top right list the names of other players. make them clickable to interact in some way... friend, chat, block?)