Working on Server Integration for all methods
I have been working on the game logic to integrate when the server should be called. Need to go over the lambda and the onDataRecieved still for call backs.

Also still having uri error (with regard to creating session?).

Next need to get it sent back...
And then get calls to the server for my other events...

Next...
1. Server integration
2 menus for game start and options
3. Spawn system for interactive points
4. refine strings and visual assets
5. Add information about the other players to the in game menu (in the top right list the names of other players. make them clickable to interact in some way... friend, chat, block?)
6. NPC AI (movement and "knowledge"/dialogue AI)