Interface for PlayerName and PlayerSprite
Added some interface to allow the player to set a sprite and name on the launch menu. when find match is pressed this will be sent to the server so it can be set... then when players make a call to set ui we need to send it to them.

Next...
1. Server integration (get two players on and data received working)
2. Spawn system for interactive points
3. refine strings and visual assets
4. menu for options on launch?
5. Add information about the other players to the in game menu (in the top right list the names of other players. make them clickable to interact in some way... friend, chat, block?)
6. NPC AI (movement and "knowledge"/dialogue AI)