Player Can Connect, Cannot Stay Connected
Currently getting player session / game session successfully. However the player session is returning with the reserved keyword.  Try to fix... as i type this i wonder if i can just set it in the lambda after it is made?  try this and fix it.

Then... we need to get onDataReceived to work... likely not working because of an issue with realtimeServer.js
Fix that file until we are getting onDataReceived calls. Then setup the game to make use of those calls....