Working on Connecting the Client
client is not connecting to server. fix this by continuing in EstablishConnectionToRealtimeServer and RealtimeClient. I need to have a Connect method in Realtime Client outside of the initialization and call it in gamemanager, I believe.

Next, test this on the server. Fix issues as needed. Then start removing unused code and testing that it is still functional until we have it as clean as possible.

Next..
1. Spawn system for interactive points
2. refine strings and visual assets. implement guess data sharing in the end guess cutscene.
3. menu for options on launch?
4. Test on server.

additional
5. NPC AI (movement and "knowledge"/dialogue AI)